Refactor agent service classes. Use factory for dynamic agents based on model opts
