+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Services/ServiceRegistrations.cs
@@ -14,14 +14,12 @@ namespace Azure.Sdk.Tools.Cli.Services
         public static void RegisterCommonServices(IServiceCollection services)
         {
             services.AddSingleton<IAzureService, AzureService>();
-            services.AddSingleton<IAIAgentService, AIAgentService>();
+            services.AddSingleton<IAzureAgentServiceFactory, AzureAgentServiceFactory>();
             services.AddSingleton<IGitHubService, GitHubService>();
             services.AddSingleton<IGitHelper, GitHelper>();
             services.AddSingleton<ITypeSpecHelper, TypeSpecHelper>();
             services.AddSingleton<IDevOpsConnection, DevOpsConnection>();
             services.AddSingleton<IDevOpsService, DevOpsService>();
         }
-
-
     }
 }
